chore: extract separate config tests
